Introduce SymptomOnsetDate screen (#761)
Why: we would like to be have the option to include the symptom onset date in the positive test result submission.

This commit:
Adds a SymptomOnsetDate screen to the AffectedUserFlow that can be activated with a build flag
The SymptomOnsetDate either collects a date or null if the user was asymptomatic. This value is saved to the context. Nothing else is done with this value. A future PR will add the symptom onset date to the network request if it exists.
Adds the react-native-static-safe-area-insets package, which enables us to get the device's insets without using a hook.