Use explict applinksdomain and update build script (#779)
Why:
Currently the universal link for enx is failing to route to the correct
app.

This commit:
Updates the applink domain in the xcodeproj capabilities to not use the
wildcard character to register the domain. That is, we switched
applinks:*.en.express for applinks:xx-xx.en.express. As a
consequence we needed to update the build scripts to set the correct
domain.

We also update the logging for the build scripts to be more readable.

Co-authored-by: artgibson <332115+artgibson@users.noreply.github.com>